LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/path-objects/2d.path.arcTo.nonfinite-expected.txt

2d.path.arcTo.nonfinite

arcTo() with Infinity/NaN is ignored


PASS arcTo() with Infinity/NaN is ignored 
